feat: implement parse test functionality with Gemini integration
- Added new repository functions for managing parse test runs and summaries.
- Created error handling for parse test operations.
- Implemented Gemini API interactions for syllabus parsing.
- Developed validation logic for syllabus candidates.
- Introduced normalization functions for parsing and processing syllabus data.
- Added prompt generation for structured syllabus extraction.
- Implemented logging for parse test activities and errors.
- Created data models for parse test results and schedules.
