+++ app/parse-test/components/page-header.tsx
@@ -1,0 +1,25 @@
+import Link from "next/link";
+import { SignOutButton } from "@/components/auth/sign-out-button";
+
+export function PageHeader({ displayName }: { displayName: string }) {
+  return (
+    <div className="flex items-center justify-between gap-4">
+      <div>
+        <p className="text-xs font-semibold uppercase tracking-[0.22em] text-zinc-500 dark:text-zinc-400">
+          TaskMaster
+        </p>
+        <h1 className="mt-2 text-4xl font-semibold tracking-tight">ParseTest</h1>
+        <p className="mt-2 text-sm text-zinc-600 dark:text-zinc-400">Signed in as {displayName}</p>
+      </div>
+      <div className="flex items-center gap-3">
+        <Link
+          href="/"
+          className="rounded-full border border-zinc-300 px-4 py-2 text-sm text-zinc-600 transition hover:border-zinc-400 hover:text-zinc-950 dark:border-zinc-700 dark:text-zinc-300 dark:hover:border-zinc-500 dark:hover:text-zinc-100"
+        >
+          Back home
+        </Link>
+        <SignOutButton />
+      </div>
+    </div>
+  );
+}
